Organize blsftcosi folders
This moves the files around to keep the same structure as more
recent package like byzcoin with the service and api at the top
level and further implementation in different package. CLI is
called blscosi for simplicity

Fixes #1562
