--- blsftcosi/server_test.go
@@ -1,11 +1,0 @@
-package main
-
-import (
-	"testing"
-
-	"github.com/dedis/onet/log"
-)
-
-func TestMain(m *testing.M) {
-	log.MainTest(m)
-}
